requirements.txt
appdirs==1.4.3
configparser==3.5.0
dj-database-url==0.4.2
Django==1.10.6
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
enum34==1.1.6
flake8==3.3.0
mccabe==0.6.1
packaging==16.8
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
PyJWT==1.4.2
pyparsing==2.2.0
python-env==1.0.0
six==1.10.0
